+++ RELEASES.md
@@ -6,7 +6,9 @@ Released on TBC.
 None
 
 ### Enhancements
-None
+- Added Binance Spot/Margin external order handling
+- Added support for millisecond bar aggregation
+- Added configurable `debug` mode for engines (with extra debug logging)
 
 ### Fixes
 None